+++ pytket/extensions/aqt/multi_zone_architecture/circuit_routing/route_circuit.py
@@ -1,11 +1,16 @@
 from pytket.circuit import Circuit
-
 from .greedy_routing import GreedyCircuitRouter
-from .partition_routing import PartitionCircuitRouter
 from .settings import RoutingSettings, RoutingAlg
 from ..architecture import MultiZoneArchitecture
 from ..circuit.helpers import ZonePlacement
 from ..circuit.multizone_circuit import MultiZoneCircuit
+from ..graph_algs.mt_kahypar_check import (
+    MT_KAHYPAR_INSTALLED,
+    MissingMtKahyparInstallError,
+)
+
+if MT_KAHYPAR_INSTALLED:
+    from .partition_routing import PartitionCircuitRouter
 
 
 def route_circuit(
@@ -19,10 +24,7 @@ def route_circuit(
      physical operations where needed
 
     The Circuit provided cannot have more qubits than allowed by
-     the architecture. If no initial placement of qubits into
-    the architecture zones is provided, the qubits will be
-     placed using an internal algorithm in a "balanced" way across
-    the available zones.
+     the architecture.
 
     :param settings: Settings used to Route Circuit
     :param circuit: A pytket Circuit to be routed
@@ -32,9 +34,12 @@ def route_circuit(
     """
     match settings.algorithm:
         case RoutingAlg.graph_partition:
-            return PartitionCircuitRouter(
-                circuit, arch, initial_placement, settings
-            ).get_routed_circuit()
+            if MT_KAHYPAR_INSTALLED:
+                return PartitionCircuitRouter(
+                    circuit, arch, initial_placement, settings
+                ).get_routed_circuit()
+            else:
+                raise MissingMtKahyparInstallError()
         case RoutingAlg.greedy:
             return GreedyCircuitRouter(
                 circuit, arch, initial_placement, settings