+++ pytket/extensions/aqt/multi_zone_architecture/graph_algs/mt_kahypar_check.py
@@ -1,0 +1,10 @@
+import importlib
+
+MT_KAHYPAR_INSTALLED = importlib.util.find_spec("mtkahypar") is not None
+"""Whether or not the mtkahypar package is available"""
+
+
+class MissingMtKahyparInstallError(Exception):
+
+    def __init__(self) -> None:
+        super().__init__("Graph partitioning requires mtkahypar package")