+++ pytket/extensions/aqt/multi_zone_architecture/initial_placement/settings.py
@@ -16,22 +16,21 @@ class InitialPlacementAlg(Enum):
     graph_partition = 2
 
 
-MIN_ZONE_FREE_SPACE: Final = 2
+MIN_ZONE_FREE_SPACE: Final = 1
 
 
 @dataclass
 class InitialPlacementSettings:
-    algorithm: InitialPlacementAlg = InitialPlacementAlg.graph_partition
+    algorithm: InitialPlacementAlg = InitialPlacementAlg.qubit_order
     zone_free_space: int = 2
     manual_placement: ZonePlacement | None = None
     n_threads: int = 1
     max_depth: int = 200
 
-    def __post_init__(self):
+    def __post_init__(self) -> None:
         if self.zone_free_space < MIN_ZONE_FREE_SPACE:
             raise InitialPlacementSettingsError(
-                f"{self.zone_free_space.__name__}"
-                f" must be larger than {MIN_ZONE_FREE_SPACE}"
+                f"zone_free_space must be larger than {MIN_ZONE_FREE_SPACE}"
             )
         if self.algorithm == InitialPlacementAlg.manual and not self.manual_placement:
             raise InitialPlacementSettingsError(