+++ docs/changelog.rst
@@ -4,8 +4,15 @@ Changelog
 Unreleased
 ----------
 
+* Add graph partition based routing for AQTMultiZoneBackend
+
+0.34.0 (August 2024)
+--------------------
+
 * Implement new aqt api
 * Fix handling of ``simplify_initial`` kwarg in ``process_circuits()``.
+* Updated Qiskit AQT Provider to v1.8.
+* Updated pytket version requirement to 1.31.
 
 0.33.0 (March 2024)
 -------------------